% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclideanDistance.R
\name{calcEDistance}
\alias{calcEDistance}
\title{Calculate Euclidean Distance}
\usage{
calcEDistance(data, projection, id, attrGroups, chem, limit)
}
\arguments{
\item{data}{selected data}

\item{projection}{groups to project against}

\item{id}{ID column}

\item{attrGroups}{selected attribute group}

\item{chem}{chemical data}

\item{limit}{number of closest matches to return}
}
\value{
data frame of closest matches
}
\description{
Calculate Euclidean Distance
}
\examples{
calcEDistance(rvals$selectedData,input$projectionGroup,input$edsampleID,rvals$attrGroups,rvals$chem,input$EDlimit)
}
