% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mahalanobis_distance.R
\name{calculate_mahalanobis_distance}
\alias{calculate_mahalanobis_distance}
\title{Calculate Squared Mahalanobis Distance}
\usage{
calculate_mahalanobis_distance(x_bar, y_bar, f0, Sx, Sy, m)
}
\arguments{
\item{x_bar}{A numeric vector representing the mean vector \eqn{\bar{x}}.}

\item{y_bar}{A numeric vector representing the mean vector \eqn{\bar{y}}.}

\item{f0}{A numeric value representing the scaling factor \eqn{f_0}.}

\item{Sx}{A numeric matrix representing the covariance matrix \eqn{S_x}.}

\item{Sy}{A numeric matrix representing the covariance matrix \eqn{S_y}.}

\item{m}{An integer representing the sample size \eqn{m}.}
}
\value{
A numeric value representing the squared Mahalanobis distance.
}
\description{
This function calculates the squared Mahalanobis distance between two mean vectors,
given the scaling factor, covariance matrices, and sample size.
}
\examples{
x_bar <- c(1, 2)
y_bar <- c(2, 3)
f0 <- 1
Sx <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)
Sy <- matrix(c(1, 0.3, 0.3, 1), nrow = 2)
m <- 10

distance <- calculate_mahalanobis_distance(x_bar, y_bar, f0, Sx, Sy, m)
print(distance)
}
