% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetData.R
\name{chooseDFServer}
\alias{chooseDFServer}
\title{Choose dataset server}
\usage{
chooseDFServer(id, rvals)
}
\arguments{
\item{id}{module ID}

\item{rvals}{reactive values object}
}
\value{
module server
}
\description{
Choose dataset server
}
\examples{
chooseDFServer("chooseDF",rvals)
}
