% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetData.R
\name{chooseDFUI}
\alias{chooseDFUI}
\title{Choose dataset UI}
\usage{
chooseDFUI(id, label = "choose DF")
}
\arguments{
\item{id}{module ID}

\item{label}{module label}
}
\value{
module
}
\description{
Choose dataset UI
}
\examples{
chooseDFUI("chooseDF")
}
