% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTab.R
\name{clusterServer}
\alias{clusterServer}
\title{Cluster Server}
\usage{
clusterServer(input, output, session, rvals, credentials, con)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{rvals}{reactive values object}

\item{credentials}{reactive values object containing authentication state}

\item{con}{database connection}
}
\value{
server
}
\description{
Cluster Server
}
\examples{
\dontrun{
clusterServer(input, output, session, rvals, credentials, con)
}
}
