% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.profile.R
\name{comp.profile}
\alias{comp.profile}
\title{comp.profile}
\usage{
comp.profile(x, groups = NULL)
}
\arguments{
\item{x}{vector of concentrations}

\item{groups}{optional grouping vector; if provided, lines are colored by group}
}
\value{
plot
}
\description{
plots concentrations
}
\examples{
\dontrun{
comp.profile(x)
}
}
