% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect}
\usage{
connect(
  db = Sys.getenv("db"),
  host = Sys.getenv("host"),
  user = Sys.getenv("user"),
  pwd = Sys.getenv("pwd"),
  port = as.integer(Sys.getenv("port"))
)
}
\arguments{
\item{db}{database name}

\item{host}{IP or URL}

\item{user}{username}

\item{pwd}{password}

\item{port}{port number}
}
\value{
database connection
}
\description{
Connect
}
\examples{
connect(db,host,user,pwd,port)
}
