% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainputTab.R
\name{dataInputServer}
\alias{dataInputServer}
\title{Data Input}
\usage{
dataInputServer(input, output, session, rvals, con, credentials)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{rvals}{reactive values object}
}
\value{
server
}
\description{
Data Input
}
\examples{
dataInputServer(input, output, session, rvals)
}
