% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoader.R
\name{dataLoader}
\alias{dataLoader}
\title{dataLoader}
\usage{
dataLoader(filename)
}
\arguments{
\item{filename}{path to the file to load}
}
\value{
data
}
\description{
dataLoader
}
\examples{
\dontrun{
dataLoader("data.csv")
}
}
