% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoader.R
\name{dataLoaderServer}
\alias{dataLoaderServer}
\title{dataLoaderServer}
\usage{
dataLoaderServer(rvals, input, output, session, credentials, con)
}
\arguments{
\item{rvals}{reactive values object}

\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{credentials}{reactive values object containing authentication state}

\item{con}{database connection}
}
\description{
dataLoaderServer
}
\examples{
\dontrun{
dataLoaderServer(rvals, input, output, session, credentials, con)
}
}
