% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclideanDistance.R
\name{euclideanDistanceSrvr}
\alias{euclideanDistanceSrvr}
\title{Euclidean Distance Server}
\usage{
euclideanDistanceSrvr(input, output, session, rvals, credentials, con)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{rvals}{reactive values object}

\item{credentials}{reactive values object containing authentication state}

\item{con}{database connection}
}
\value{
server
}
\description{
Euclidean Distance Server
}
\examples{
\dontrun{
euclideanDistanceSrvr(input, output, session, rvals, credentials, con)
}
}
