% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclideanDistance.R
\name{euclideanDistanceTab}
\alias{euclideanDistanceTab}
\title{Euclidean Distance UI}
\usage{
euclideanDistanceTab()
}
\value{
UI
}
\description{
Euclidean Distance UI
}
\examples{
euclideanDistanceTab()
}
