% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreTab.R
\name{exploreServer}
\alias{exploreServer}
\title{Explore Server}
\usage{
exploreServer(input, output, session, rvals, con, credentials)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{rvals}{reactive values object}

\item{con}{database connection}

\item{credentials}{reactive values object containing authentication state}
}
\value{
server
}
\description{
Explore Server
}
\examples{
\dontrun{
exploreServer(input, output, session, rvals, con, credentials)
}
}
