% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_probs.R
\name{getBestGroup}
\alias{getBestGroup}
\title{Find best group based on largest values}
\usage{
getBestGroup(probsAll, eligible, method)
}
\arguments{
\item{probsAll}{result from group probability}

\item{eligible}{groups}

\item{method}{selection strategy used to choose the best group}
}
\value{
vector of group names
}
\description{
Find best group based on largest values
}
\examples{
\dontrun{
getBestGroup(probsAll, eligible, method = "max")
}
}
