% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cda.R
\name{getCDA}
\alias{getCDA}
\title{Get Canonical Discriminant Analysis Results}
\usage{
getCDA(
  df = rvals$selectedData,
  chem = rvals$chem,
  attrGroups = rvals$attrGroups
)
}
\arguments{
\item{df}{data frame}

\item{chem}{chemical element columns}

\item{attrGroups}{selected attribute}
}
\value{
list of results
}
\description{
Get Canonical Discriminant Analysis Results
}
\examples{
getCDA(df,chem,attrGroups)
}
