% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_probs.R
\name{getEligible}
\alias{getEligible}
\title{Calculate eligible groups}
\usage{
getEligible(elements, attrs, group)
}
\arguments{
\item{elements}{dataframe with chemical elements}

\item{attrs}{dataframe with descriptive attributes}

\item{group}{selected group attribute}
}
\value{
vector of eligible groups for further analysis
}
\description{
number of rows must be 2 more than the number of columns per group
}
