% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{getLDA}
\alias{getLDA}
\title{Compute LDA}
\usage{
getLDA(
  df = rvals$selectedData,
  chem = rvals$chem,
  attrGroups = rvals$attrGroups
)
}
\arguments{
\item{df}{data frame}

\item{chem}{chemical element columns}

\item{attrGroups}{selected attribute}
}
\value{
list of results
}
\description{
Compute LDA
}
\examples{
getLDA(df,chem,attrGroups)
}
