% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_probs.R
\name{getMahalanobis}
\alias{getMahalanobis}
\title{Get Mahalanobis distance}
\usage{
getMahalanobis(row, data)
}
\arguments{
\item{row}{to get distance for}

\item{data}{dataset to compare against}
}
\value{
distance
}
\description{
Get Mahalanobis distance
}
\examples{
getMahalanobis(data[r,],data[,data[[group]] == grp])
}
