% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_probs.R
\name{group.mem.probs}
\alias{group.mem.probs}
\title{Function for calculating group membership probabilities by chemical compositional distance using Mahalanobis distances and Hotellings T^2 statistic}
\usage{
group.mem.probs(data, chem, group, eligible, method = "Hotellings", ID)
}
\arguments{
\item{data}{dataframe from imported data}

\item{chem}{names of columns with chemical elements}

\item{group}{selected group attribute}

\item{eligible}{groups to include in assignment}

\item{method}{distance/probability method}

\item{ID}{identifier column}
}
\value{
data frame
}
\description{
Function for calculating group membership probabilities by chemical compositional distance using Mahalanobis distances and Hotellings T^2 statistic
}
\examples{
\dontrun{
group.mem.probs(data, chem, group, eligible, method = "Hotellings", ID = "rowid")
}
}
