% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupMembership.R
\name{groupTab}
\alias{groupTab}
\title{UI elements for group membership tab}
\usage{
groupTab()
}
\value{
UI
}
\description{
UI elements for group membership tab
}
\examples{
groupTab()
}
