% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{load_elements}
\alias{load_elements}
\title{Load Elements from CSV}
\usage{
load_elements()
}
\value{
A data frame containing the elements defined in
\code{elements.csv}.
}
\description{
This function loads the \code{elements.csv} file included with the
\code{ArchaeoDash} package and returns it as a data frame.
}
\details{
The function uses \code{system.file()} to locate the \code{elements.csv}
file in the \code{app} directory of the installed package.
}
\examples{
\dontrun{
elements <- load_elements()
head(elements)
}

}
