% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginServer.R
\name{loginServer}
\alias{loginServer}
\title{loginServer}
\usage{
loginServer(
  con,
  input = input,
  output = output,
  session = session,
  credentials = credentials
)
}
\arguments{
\item{con}{Database connection}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{credentials}{reactive values object}
}
\value{
credentials object
}
\description{
loginServer
}
\examples{
loginServer(con,input,output,session,credentials)
}
