% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot function}
\usage{
multiplot(selectedData, attrGroups, xvar, yvar, ptsize, interactive = F)
}
\arguments{
\item{selectedData}{data frame for plotting}

\item{attrGroups}{attribute group}

\item{xvar}{x variable}

\item{yvar}{y variable}

\item{ptsize}{point size}

\item{interactive}{whether to return plotly object or ggplot object}
}
\value{
plot object
}
\description{
Multiplot function
}
\examples{
multiplot(selectedData = rvals$selectedData,attrGroups = rvals$attrGroups,xvar  = input$xvar2, yvar = input$yvar2,ptsize = input$ptsize, interactive = input$interactive)
}
