% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietly.R
\name{mynotification}
\alias{mynotification}
\title{custom notification}
\usage{
mynotification(
  message,
  type = c("default", "message", "warning", "error", "success"),
  duration = 5
)
}
\arguments{
\item{message}{message to display}

\item{type}{type of message}
}
\value{
notification
}
\description{
custom notification
}
\examples{
mynotification("hello world")
}
