% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinationTab.R
\name{ordinationServer}
\alias{ordinationServer}
\title{Ordination Server}
\usage{
ordinationServer(input, output, session, rvals)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{rvals}{reactive values object}
}
\value{
server
}
\description{
Ordination Server
}
\examples{
ordinationServer(input,output,session,rvals)
}
