% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot Missing Data}
\usage{
plot_missing(data)
}
\arguments{
\item{data}{elements}
}
\value{
ggplot object
}
\description{
From DataExplorer
}
\examples{
plot_missing(data)
}
