% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{plotLDAvectors}
\alias{plotLDAvectors}
\title{plot LDA vectors}
\usage{
plotLDAvectors(lda_model)
}
\arguments{
\item{lda_model}{lda model}
}
\value{
plot
}
\description{
plot LDA vectors
}
\examples{
plotLDAvectors(lda_model)
}
