% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{profile_missing}
\alias{profile_missing}
\title{Profile Missing Data}
\usage{
profile_missing(data)
}
\arguments{
\item{data}{data to evalue}
}
\value{
data frame
}
\description{
From DataExplorer
}
\examples{
profile_missing(data)
}
