% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietly.R
\name{quietly}
\alias{quietly}
\title{Quiet Warning Function}
\usage{
quietly(.expr, label = NULL)
}
\arguments{
\item{.expr}{expression}

\item{label}{optional label shown when reporting suppressed warnings}
}
\value{
results of function
}
\description{
Quiet Warning Function
}
\examples{
\dontrun{
f = sum(1 + 'a')
quietly(f())
}
}
