% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{replaceCell}
\alias{replaceCell}
\title{Replace values}
\usage{
replaceCell(rowid, col, value, rvals, con, credentials, input, output, session)
}
\arguments{
\item{rowid}{rowid of value to change}

\item{col}{column of value to change}

\item{value}{new value}

\item{rvals}{reactive values object}

\item{con}{database connection}

\item{credentials}{reactive values object containing authentication state}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
reactive values object
}
\description{
Replace values
}
\examples{
\dontrun{
replaceCell(rowid, col, value, rvals, con, credentials, input, output, session)
}
}
