% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runArchaeoDash}
\alias{runArchaeoDash}
\title{Run Application}
\usage{
runArchaeoDash(verbose = FALSE)
}
\arguments{
\item{verbose}{logical; if TRUE, enables verbose console logging for debugging.}
}
\value{
shiny app
}
\description{
Run Application
}
\examples{
\dontrun{
runArchaeoDash()
}
}
