% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveexportTab.R
\name{saveExportServer}
\alias{saveExportServer}
\title{Save and Export Server}
\usage{
saveExportServer(input, output, session, rvals)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{rvals}{reactive values object}
}
\value{
server
}
\description{
Save and Export Server
}
\examples{
saveExportServer(input,output,session,rvals)
}
