% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetData.R
\name{subsetDataServer}
\alias{subsetDataServer}
\title{Subset data server}
\usage{
subsetDataServer(input, output, session, rvals)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{rvals}{reactive values object}
}
\value{
server
}
\description{
Subset data server
}
\examples{
subsetDataServer(input,output,session,rvals)
}
