% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetData.R
\name{subsetModServer}
\alias{subsetModServer}
\title{Subset Module Server}
\usage{
subsetModServer(id, rvals)
}
\arguments{
\item{id}{module ID}

\item{rvals}{reactive values object}
}
\value{
module server
}
\description{
Subset Module Server
}
\examples{
subsetModServer("subset",rvals)
}
