% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetData.R
\name{subsetModUI}
\alias{subsetModUI}
\title{Subset Module UI}
\usage{
subsetModUI(id, label = "subset")
}
\arguments{
\item{id}{module ID}

\item{label}{module label}
}
\value{
module
}
\description{
Subset Module UI
}
\examples{
subsetModUI("subset")
}
