% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCurrent.R
\name{updateCurrent}
\alias{updateCurrent}
\title{Save current table to database}
\usage{
updateCurrent(rvals, con, credentials, input, output, session)
}
\arguments{
\item{rvals}{reactive values object}

\item{con}{database connection}

\item{credentials}{reactive values object}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\description{
Save current table to database
}
\examples{
updateCurrent(rvals,con,credentials,input,output,session)
}
