% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeassignTab.R
\name{visualizeAssignServer}
\alias{visualizeAssignServer}
\title{Visualize Assign Server}
\usage{
visualizeAssignServer(input, output, session, rvals, credentials, con)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{rvals}{reactive values object}
}
\value{
server
}
\description{
Visualize Assign Server
}
\examples{
visualizeAssignServer(input,output,session,rvals)
}
