% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScore.R
\name{zScore}
\alias{zScore}
\title{Convert to z score}
\usage{
zScore(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
data frame of z scores
}
\description{
Convert to z score
}
\examples{
zScore(x)
}
