% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.profile.R
\name{comp.profile}
\alias{comp.profile}
\title{comp.profile}
\usage{
comp.profile(x)
}
\arguments{
\item{x}{vector of concentrations}
}
\value{
plot
}
\description{
plots concentrations
}
\examples{
comp.profile(x)
}
