% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoader.R
\name{dataLoaderServer}
\alias{dataLoaderServer}
\title{dataLoaderServer}
\usage{
dataLoaderServer(rvals, input, output, session, credentials, con)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
dataLoaderServer
}
\examples{
dataLoaderServer(input,output,session, credentials, con)
}
