% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreTab.R
\name{exploreServer}
\alias{exploreServer}
\title{Explore Server}
\usage{
exploreServer(input, output, session, rvals)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{rvals}{reactive values object}
}
\value{
server
}
\description{
Explore Server
}
\examples{
exploreServer(input,output,session,rvals)
}
