% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_probs.R
\name{getEligible}
\alias{getEligible}
\title{Calculate eligible groups}
\usage{
getEligible(data, chem, group)
}
\arguments{
\item{data}{dataframe from imported data}

\item{chem}{names of columns with chemical elements}

\item{group}{selected group attribute}
}
\value{
vector of eligible groups for further analysis
}
\description{
number of rows must be 2 more than the number of columns per group
}
\examples{
getEligible(data,chem,group)
}
