% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_probs.R
\name{group.mem.probs}
\alias{group.mem.probs}
\title{Function for calculating group membership probabilities by chemical compositional distance using Mahalanobis distances and Hotellings T^2 statistic}
\usage{
group.mem.probs(data, chem, group, eligible, method = "Hotellings", ID)
}
\arguments{
\item{elements}{transformed element data}

\item{assigned}{group designation by sample}
}
\value{
data frame
}
\description{
Function for calculating group membership probabilities by chemical compositional distance using Mahalanobis distances and Hotellings T^2 statistic
}
\examples{
group.mem.probs(elements,assigned)
}
