% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupMembership.R
\name{groupServer}
\alias{groupServer}
\title{Group Membership Server}
\usage{
groupServer(input, output, session, rvals, credentials, con)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{rvals}{reactive values object}
}
\value{
server
}
\description{
Group Membership Server
}
\examples{
groupServer(input,output,session,rvals)
}
