% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mainPlot}
\alias{mainPlot}
\title{Main plotting function}
\usage{
mainPlot(plotdf, xvar, yvar, attrGroups, Conf, int.set, theme = "viridis")
}
\arguments{
\item{plotdf}{data frame for plotting}

\item{xvar}{x variable}

\item{yvar}{y variable}

\item{attrGroups}{attribute group}

\item{Conf}{whether to draw confidence ellipses}

\item{int.set}{confidence level}

\item{theme}{color theme}
}
\value{
plotly object
}
\description{
Main plotting function
}
\examples{
mainPlot(plotdf = rvals$plotdf,xvar = input$xvar,yvar = input$yvar,attrGroups = rvals$attrGroups,Conf = input$Conf, int.set = input$int.set)
}
