% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{replaceCell}
\alias{replaceCell}
\title{Replace values}
\usage{
replaceCell(rowid, col, value, rvals, con, credentials, input, output, session)
}
\arguments{
\item{rowid}{rowid of value to change}

\item{col}{column of value to change}

\item{value}{new value}

\item{rvals}{reactive values object}
}
\value{
reactive values object
}
\description{
Replace values
}
\examples{
replaceCell(rvals,rowid,col,value)
}
