% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{replaceCell}
\alias{replaceCell}
\title{Replace values}
\usage{
replaceCell(rvals, rowid, col, value)
}
\arguments{
\item{rvals}{reactive values object}

\item{rowid}{rowid of value to change}

\item{col}{column of value to change}

\item{value}{new value}
}
\value{
reactive values object
}
\description{
Replace values
}
\examples{
replaceCell(rvals,rowid,col,value)
}
